Add files via upload
Exercise 18: Temperature Converter

You'll often need to determine which part of a program is run based on user input or other events.

Create a program that converts temperatures from Fahrenheit to Celcius or from Celsius to Fahrenheit.
Prompt for the starting temperature.  The program should prompt for the type of conversion and then
perform the conversion.