Add files via upload
Exercise 19: BMI Calculator
Create a program to calculate the body mass index (BMI) for a person using the
person's height in inches and weight in pounds.  The program should prompt
the user for weight and height.

Calculate the BMI by usingthe following formula:

bmi = (weight / (height * height)) * 703

If the BMI is between 18.5 and 25, display that the person is at a normal weight.
If they are out of that range, tell them if they are underweight or overweight
and tell them to consult their doctor.