Exercise 24: Anagram Checker
Exercise 24: Anagram Checker

Create a program that compares two strings and determines if the two strings are
anagrams.  The program should prompt for both input strings and display the output
as shown in the example that follows.

Constraints:
Implement the program using a function called 'isAnagram', which takes two words as
its arguments and returns true or false.  Invoke this function from your main
program.

Check that both words are the same length