Exercise 32: Guess the Number Game
Exercise 32: Guess the Number Game

Prompt for the difficulty level, and then start the game.  The computer picks a random number in that 
range and prompts the player to guess that number.  Each time the playher guesses, the computer should
give the player a hint as to whether the number is too high or too low.  The computer should also keep 
track of the number of guesses.  Once the player guesses the correct number, the computer should 
preseent the number of guesses and ask if the player would like to play again.

Constraints:
Don't allow any non-numeric data entry.
During the game, count non-numeric entries as wrong guesses.