Exercise 09: Paint Calculator
Exercise 9: Paint Calculator

Calculate the gallons of paint needed to paint the ceiling of a room.  Prompt for the length and width, and assume one gallon covers 350 square feet.  Display the number of gallons needed to paint the ceiling as a whole number.

Constraints:
Use a constant to hold the conversion rate.
Ensure that you round up to the next whole gallon.