Execise 16: Legal Driving Age
Exercise 16: Legal Driving Age

Write a program that asks the user for their age and
compare it to the legal driving age of sixteen.  If the
user is sixteen or older, then the program should
display "You are old enough to legally drive."  If the
user is under sixteen, the program should display
"You are not old enough to legally drive.