Exercise 08:  Pizza Party
Exercise 08:  Pizza Party

Write a program to evenly divide pizzas.  Prompt for the number of people, the number of pizzas, 
and the number of slices per pizza.  Ensure that the number of pieces comes out even.  Display 
the number of pieces of pizza each person should get.  If there are leftovers, show the number 
of leftover pieces.