Exercise 22: Comparing Numbers
Exercise 22: Comparing Numbers

Write a program that asks for three numbers.  Check first to see that all
numbers are different.  If they're not different, then exit the program.
Otherwise, display the largest number of the three.

Constraint:
Write the algorithm manually.  Don't use a built-in function for finding the
largest number in a list.