Exercise 34: Employee List Removal
Exercise 34: Employee List Removal

Create a small program that contains a list of employee names.  Print out the
list of names when the program runs the first time.  Prompt the employee name
and remove that specific name from the list of names.  Display the remaining
employees, each on its own line.

Constraints:
Use an array or list to store the names.